requirements.txt
absl-py==1.2.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.2.0
certifi
charset-normalizer==2.1.1
click==8.1.3
croniter==1.3.15
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.1
dill==0.3.6
docker-pycreds==0.4.0
dominate==2.7.0
efficientnet-pytorch==0.7.1
einops==0.4.1
exceptiongroup==1.1.2
fastapi==0.88.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
h11==0.14.0
idna==3.3
imageio==2.21.2
importlib-metadata==4.12.0
inquirer==2.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lightning==1.9.5
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lmdb==1.3.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
multidict==6.0.4
natsort==8.2.0
networkx==2.6.3
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.19.4
psutil==5.9.2
pthflops==0.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
rawpy==0.18.1
readchar==4.0.5
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.22.0
starsessions==1.3.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.7
torch==1.13.1
torchaudio==0.7.0a0+a853dff
torchmetrics==0.11.4
torchvision==0.8.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions
urllib3==1.26.12
uvicorn==0.22.0
wandb==0.13.2
warmup-scheduler==0.3
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.2.2
yarl==1.9.2
zipp==3.8.1
